nlp-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
tweepy==4.14.0
facebook-sdk==3.1.0
google-api-python-client==2.110.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2
python-dotenv==1.0.0
textblob==0.17.1
vaderSentiment==3.3.2